requirements.txt
torch==1.9.0
numpy==1.19.5
Pillow==8.3.1
streamlit==0.87.0
